+++ src/measurement/decoding/plainHandler.js
@@ -1,12 +1,12 @@
 'use strict';
 
+const log = require('../../log');
+
 // this handler is only for debugging..
 
 module.exports = {
   decodeMessage: function (message) {
-    /* eslint-disable no-console */
-    console.log(message);
-    /* eslint-enable no-console */
+    log.debug(message);
 
     return Promise.resolve(message);
   }