+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # @sensebox/opensensemap-api-models Changelog
 
 ## Unreleased
+- Re-add bunyan logging
 
 ## v0.0.1
 - Initial Release after splitting api and models