+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Re-add bunyan logging
+- toLowerCase email parameter in User `initPasswordReset` to conform to rest of email address handling
 
 ## v0.0.1
 - Initial Release after splitting api and models