+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Validate that boxes need at least one sensor
 - Move measurement deletion code from box to sensor
 - Move findLastMeasurement from measurement to sensor
+- Pin dependency mongoose to version 4.11.11
 
 ## v0.0.1
 - Initial Release after splitting api and models