+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Move findLastMeasurement from measurement to sensor
 - Pin dependency mongoose to version 4.11.11
 - Use `json` as default for `format` parameter of `Box.findBoxesLastMeasurements`
+- Parse timestamps based on RFC 3339
 
 ## v0.0.1
 - Initial Release after splitting api and models