+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Pin dependency mongoose to version 4.11.11
 - Use `json` as default for `format` parameter of `Box.findBoxesLastMeasurements`
 - Parse timestamps based on RFC 3339
+- Convert RFC 3339 nanoseconds to milliseconds
 
 ## v0.0.1
 - Initial Release after splitting api and models