+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - Validate that boxes need at least one sensor
 - Move measurement deletion code from box to sensor
 - Move findLastMeasurement from measurement to sensor
-- Pin dependency mongoose to version 4.11.11
+- Pin dependency mongoose to version 4.13.6
 - Use `json` as default for `format` parameter of `Box.findBoxesLastMeasurements`
 - Parse timestamps based on RFC 3339
 - Convert RFC 3339 nanoseconds to milliseconds